Bugfix (#115)
* Dodatne izmene u kalkulacijama, centrima i RSD umesto DIN (#105)


* Fixevi (#108)

* Popravljeno sve sto je Milena trazila.
Ostaje da se testira Dacicev rad za Obracun komponentu.

* Popravljeno sve sto je Milena trazila.
Ostaje da se testira Dacicev rad za Obracun komponentu.

* Gotovo

* Dodatne izmene u kalkulacijama, centrima i RSD umesto DIN (#105) (#106)



* Popravljeno sve sto je Milena trazila.
Ostaje da se testira Dacicev rad za Obracun komponentu.

* Gotovo

* Popravljeno sve sto je Milena trazila.
Ostaje da se testira Dacicev rad za Obracun komponentu.

* Gotovo


* manage-users-permission-fix: dodate nove permisije za edit, uvedena p… (#110)

* Dodatne izmene u kalkulacijama, centrima i RSD umesto DIN (#105) (#106)



* manage-users-permission-fix: dodate nove permisije za edit, uvedena preduzeca za user-a


* manage-usrs-bugfix: pokriven slucaj restrikcije brisanja admin-a (#114)

+++ src/app/components/manage-users/manage-users.component.ts
@@ -5,6 +5,9 @@ import { ManageUsersService } from '../../services/manage-users/manage-users.ser
 import { Authority } from 'src/app/shared/enums/permissions';
 import {Company} from "../../shared/invoice.model";
 
+import {UserService} from "../../services/login/user.service";
+
+
 @Component({
   selector: 'app-manage-users',
   templateUrl: './manage-users.component.html',
@@ -13,6 +16,7 @@ import {Company} from "../../shared/invoice.model";
 export class ManageUsersComponent implements OnInit {
   userEditForm: FormGroup;
   userAddForm: FormGroup;
+  loggedUser: User;
 
   preduzeca: Company[] = [];
   userToEdit: User | undefined;
@@ -25,7 +29,8 @@ export class ManageUsersComponent implements OnInit {
 
   constructor(
     private formBuilder: FormBuilder,
-    private serviceBack: ManageUsersService
+    private serviceBack: ManageUsersService,
+    private userService: UserService
   ) {
     this.userEditForm = this.formBuilder.group({
       username: ['', [Validators.required, Validators.minLength(3)]],
@@ -63,6 +68,11 @@ export class ManageUsersComponent implements OnInit {
     this.serviceBack.svaPreduzeca().subscribe((preduzeca) => {
       this.preduzeca = preduzeca;
     });
+
+    this.userService.getLoggedInUser().subscribe((user) => {
+      this.loggedUser = user
+    });
+
     this.serviceBack.listAllUsers().subscribe(
       (res) => {
         this.users = res;
@@ -92,6 +102,16 @@ export class ManageUsersComponent implements OnInit {
   }
 
   delete(user: User) {
+    for (let i = 0; i < user.permissions.length; i++) {
+      if(user.permissions[i].name === Authority.ADMIN){
+        alert("Ne mosete da brisete admin-a");
+        return;
+      }
+    }
+    if(this.loggedUser.userId === user.userId){
+      alert("Ne smete da brisete samog sebe");
+      return;
+    }
     this.serviceBack.deleteUser(user.userId).subscribe(
       (res) => {
         this.hiddenEdit = true;
